requirements.txt
alembic==1.13.2
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
BareNecessities==0.2.8
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
Brotli==1.1.0
cachelib==0.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
cryptography==44.0.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dnspython==2.7.0
docopt==0.6.2
dominate==2.9.1
email_validator==2.2.0
executing==2.1.0
fastjsonschema==2.20.0
Flask==3.0.3
Flask-Assets==2.1.0
Flask-Caching==2.3.0
Flask-Compress==1.17
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
future==1.0.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
humanfriendly==10.0
idna==3.10
imageio==2.35.1
importlib_resources==6.4.4
iniconfig==2.0.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lazy_loader==0.4
limits==3.13.0
llvmlite==0.43.0
loguru==0.7.3
lxml==5.3.0
lxml_html_clean==0.3.1
Mail==2.1.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
MouseInfo==0.1.3
mpmath==1.3.0
msgspec==0.18.6
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
numba==0.60.0
numpy==2.0.2
oauthlib==3.2.2
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
prompt_toolkit==3.0.48
protobuf==5.28.2
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyMatting==1.1.12
PyMsgBox==1.0.9
pyperclip==1.9.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytweening==1.2.0
pytz==2024.2
# pywin32==308
pyzmq==26.2.0
rcssmin==1.1.3
redis==5.0.8
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.32
stack-data==0.6.3
StrEnum==0.4.15
sympy==1.13.3
tifffile==2024.9.20
tinycss2==1.4.0
token-bucket==0.3.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
unidecode==1.3.8
urllib3==2.2.2
visitor==0.1.3
wcwidth==0.2.13
webassets==2.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
yarg==0.1.9
zstandard==0.23.0
